+++ ipipneo/__init__.py
@@ -1,6 +1,6 @@
 """Import of the classes to be used."""
 
 __name__ = "five-factor-e"
-__version__ = "1.11.0"
+__version__ = "1.12.0"
 
 from ipipneo.ipipneo import IpipNeo